feat(config): Fix issue with /static/ image files with '@' characters replacing with '_' with used.
When testing the `/static/iguana-8084900@5257x3505.jpg` on disk gets renamed to `/static/iguana-8084900_5257x3505.jpg` when used within the HTML content. It appears the `@` character gets converted over to a `_`. I bet this is because of the URL when trying ot access this image.

For now the logic I have removes this image from the course because it recognizes that it's not used. We'll have to use a search for disk `/static/<filename>` in a lookup for the `/policies/assets.json` file for the `displayName` property, then retrieve the dictionary key associated with that property and use that name when searching the course to see if that image is used or not.