feat: Run multiprocessing for this application and create separate loggers for application and courses.
**Multiprocessing**
We can adjust the following configuration values to adjust how we process multiple courses at once. For now we're keeping the number of processes low to see how this works on in production. If the chunk size is larger than the number of working processes, then we want to set the chunk size to match the number of working processes to avoid taxing the server resources.
- `NUM_COURSE_OPTIMIZATION_CHUNKS = 2`  # Specify the number of courses to optimize at a time
- `NUM_WORKER_PROCESSES = 2`  # Specify the number of worker processes

**Loggers**
Created separate loggers for the application and courses. This helps keep track at a high-level where the application is and also detailed view per course.
- `application`: this is high-level what the script is processing for a course level. output is streamed to stdout and separate `./logs/application.log` file.
- `course`: individual logs are generate for per tar.gz file based on course name at `./logs/course.<course_id>.log`.